# internal/web/handlers/handlers.go
package handlers

import (
	"context"
	"crypto/subtle"
	"log/slog"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/jackc/pgx/v5/pgxpool"

	"github.com/preshotcome/dokaz/internal/account"
	"github.com/preshotcome/dokaz/internal/analytics"
	"github.com/preshotcome/dokaz/internal/apikey"
	"github.com/preshotcome/dokaz/internal/audit"
	"github.com/preshotcome/dokaz/internal/auth"
	"github.com/preshotcome/dokaz/internal/billing"
	"github.com/preshotcome/dokaz/internal/compliance"
	"github.com/preshotcome/dokaz/internal/drill"
	"github.com/preshotcome/dokaz/internal/email"
	"github.com/preshotcome/dokaz/internal/evidence"
	"github.com/preshotcome/dokaz/internal/flags"
	"github.com/preshotcome/dokaz/internal/heartbeat"
	"github.com/preshotcome/dokaz/internal/mobileauth"
	"github.com/preshotcome/dokaz/internal/oauth"
	"github.com/preshotcome/dokaz/internal/obs"
	"github.com/preshotcome/dokaz/internal/push"
	"github.com/preshotcome/dokaz/internal/ratelimit"
	"github.com/preshotcome/dokaz/internal/web/csrf"
	"github.com/preshotcome/dokaz/internal/web/templates"
	"github.com/preshotcome/dokaz/internal/webhooks"
)

type Handlers struct {
	pool            *pgxpool.Pool
	sessions        *auth.Store
	audit           *audit.Logger
	drills          *drill.Store
	orch            *drill.Orchestrator
	heartbeats      *heartbeat.Store
	heartbeatNotify heartbeat.Notifier
	pingLimiter     *ratelimit.Limiter
	accounts        *account.Store
	billing         billing.Service
	throttle        *auth.LoginThrottle
	webhooks        *webhooks.Store
	webhookDispatch *webhooks.Dispatcher
	csrf            *csrf.Protector
	authLimiter     *ratelimit.Limiter
	appLimiter      *ratelimit.Limiter
	evidence        *evidence.Service
	signer          *evidence.Signer
	exporter        *compliance.Exporter
	purger          *compliance.Purger
	inserter        drill.RiverInserter
	obs             *obs.Provider

	mailer               *email.Mailer
	analytics            analytics.Analytics
	flags                flags.Flags
	postmarkWebhookToken string
	staffEmails          map[string]bool
	metricsToken         string
	apiKeys              *apikey.Store
	mobileTokens         *mobileauth.Store
	pushDevices          *push.Store
	v1Limiter            *ratelimit.Limiter
	sourceDir            string
	oauth                *oauth.Registry
	secureCookies        bool
	priceStarterLabel    string
	priceProLabel        string
	baseURL              string
}

type Deps struct {
	Pool            *pgxpool.Pool
	Sessions        *auth.Store
	Audit           *audit.Logger
	Drills          *drill.Store
	Orchestrator    *drill.Orchestrator
	Heartbeats      *heartbeat.Store
	HeartbeatNotify heartbeat.Notifier
	PingLimiter     *ratelimit.Limiter
	Accounts        *account.Store
	Billing         billing.Service
	Throttle        *auth.LoginThrottle
	Webhooks        *webhooks.Store
	WebhookDispatch *webhooks.Dispatcher
	CSRF            *csrf.Protector
	AuthLimiter     *ratelimit.Limiter
	AppLimiter      *ratelimit.Limiter
	Evidence        *evidence.Service
	Signer          *evidence.Signer
	Exporter        *compliance.Exporter
	Purger          *compliance.Purger
	Inserter        drill.RiverInserter
	Obs             *obs.Provider

	Mailer               *email.Mailer
	Analytics            analytics.Analytics
	Flags                flags.Flags
	PostmarkWebhookToken string
	StaffEmails          []string
	MetricsToken         string
	APIKeys              *apikey.Store
	MobileTokens         *mobileauth.Store
	PushDevices          *push.Store
	V1Limiter            *ratelimit.Limiter
	SourceDir            string
	OAuth                *oauth.Registry
	SecureCookies        bool
	PriceStarterLabel    string
	PriceProLabel        string
	BaseURL              string
}

func New(d Deps) *Handlers {
	staff := make(map[string]bool, len(d.StaffEmails))
	for _, e := range d.StaffEmails {
		staff[e] = true
	}
	return &Handlers{
		pool:            d.Pool,
		sessions:        d.Sessions,
		audit:           d.Audit,
		drills:          d.Drills,
		orch:            d.Orchestrator,
		heartbeats:      d.Heartbeats,
		heartbeatNotify: d.HeartbeatNotify,
		pingLimiter:     d.PingLimiter,
		accounts:        d.Accounts,
		billing:         d.Billing,
		throttle:        d.Throttle,
		webhooks:        d.Webhooks,
		webhookDispatch: d.WebhookDispatch,
		csrf:            d.CSRF,
		authLimiter:     d.AuthLimiter,
		appLimiter:      d.AppLimiter,
		evidence:        d.Evidence,
		signer:          d.Signer,
		exporter:        d.Exporter,
		purger:          d.Purger,
		inserter:        d.Inserter,
		obs:             d.Obs,

		mailer:               d.Mailer,
		analytics:            d.Analytics,
		flags:                d.Flags,
		postmarkWebhookToken: d.PostmarkWebhookToken,
		staffEmails:          staff,
		metricsToken:         d.MetricsToken,
		apiKeys:              d.APIKeys,
		mobileTokens:         d.MobileTokens,
		pushDevices:          d.PushDevices,
		v1Limiter:            d.V1Limiter,
		sourceDir:            d.SourceDir,
		oauth:                d.OAuth,
		secureCookies:        d.SecureCookies,
		priceStarterLabel:    d.PriceStarterLabel,
		priceProLabel:        d.PriceProLabel,
		baseURL:              d.BaseURL,
	}
}

// logger returns the package-default slog logger. Wrapped in a method so
// future per-request loggers can replace it without touching every handler.
func (h *Handlers) logger() *slog.Logger { return slog.Default() }

// layoutCtx assembles the nav chrome state for a request: the user, their
// current account + role, and the full account list for the switcher.
func (h *Handlers) layoutCtx(r *http.Request) templates.LayoutCtx {
	u, _ := auth.FromContext(r.Context())
	acct, _ := auth.CurrentAccountFromContext(r.Context())
	m, _ := auth.MembershipFromContext(r.Context())
	lc := templates.LayoutCtx{User: u, Account: acct, Membership: m}
	if imp, ok := auth.ImpersonationFromContext(r.Context()); ok {
		lc.Impersonation = imp
	}
	if u != nil {
		accts, _ := h.accounts.ListAccountsForUser(r.Context(), u.ID)
		for _, a := range accts {
			lc.Accounts = append(lc.Accounts, templates.AccountChoice{
				ID: a.ID.String(), Name: a.Name,
			})
		}
	}
	return lc
}

// requireUnlapsedTrial blocks resource-write actions for a trial account
// whose 14-day window has closed. The account stays fully readable — it
// just has to subscribe before it can create or run anything again.
func (h *Handlers) requireUnlapsedTrial(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		if acct, ok := auth.CurrentAccountFromContext(r.Context()); ok && account.TrialLapsed(*acct) {
			w.WriteHeader(http.StatusPaymentRequired)
			render(w, r, templates.TrialEnded(h.layoutCtx(r)))
			return
		}
		next.ServeHTTP(w, r)
	})
}

func (h *Handlers) Router(staticFS http.FileSystem) http.Handler {
	r := chi.NewRouter()
	r.Use(middleware.RequestID)
	r.Use(middleware.RealIP)
	r.Use(obs.TracingMiddleware)         // open the server span first
	r.Use(obs.RequestLogger(h.logger())) // one structured line per request
	r.Use(h.obs.Metrics.Middleware)      // request count + latency
	r.Use(h.obs.Recoverer)               // panic → error reporter → 500
	r.Use(securityHeaders)
	r.Use(limitRequestBody)
	r.Use(h.sessions.LoadUser)
	r.Use(h.sessions.LoadCurrentAccount(h.accounts))
	r.Use(stampAccountForLogs) // enrich the request log with account_id
	r.Use(h.csrf.Middleware)

	// Liveness: the process is up. Readiness: dependencies are reachable.
	r.Get("/healthz", func(w http.ResponseWriter, _ *http.Request) {
		w.WriteHeader(http.StatusOK)
		_, _ = w.Write([]byte("ok"))
	})
	r.Get("/readyz", obs.ReadinessHandler(map[string]func(context.Context) error{
		"database": h.pool.Ping,
	}))
	r.Handle("/metrics", h.metricsHandler())
	r.Get("/robots.txt", h.robotsTxt)

	// Public verification surface: the evidence signing keys, served as PEM
	// so anyone holding a Dokaz-issued PDF can fetch the key and verify the
	// detached signature without trusting Dokaz's own tooling.
	r.Get("/.well-known/evidence-signing-keys.pem", h.evidenceSigningKeys)

	// Inbound Postmark bounce/complaint webhook — authenticated by the
	// token path segment, CSRF-exempt (see csrf.New in main).
	r.Post("/webhooks/postmark/{token}", h.postmarkBounce)

	// Inbound Stripe webhook — authenticated by the Stripe-Signature header,
	// CSRF-exempt (the /webhooks/ prefix is exempted in csrf.New).
	r.Post("/webhooks/stripe", h.stripeWebhook)

	// Public heartbeat ping ingest — the token in the path is the only
	// credential. CSRF-exempt (the /ping/ prefix is exempted in csrf.New) and
	// rate-limited per source IP. GET and POST both accepted so a bare curl
	// works. Mounted at the top level, outside the session-gated group.
	r.Group(func(r chi.Router) {
		r.Use(h.pingLimiter.Middleware(clientIPKey))
		r.HandleFunc("/ping/{token}", h.ping)
		r.HandleFunc("/ping/{token}/{action}", h.ping)
	})

	// The versioned JSON API: API-key auth, no session/CSRF (csrf.New
	// exempts the /v1/ prefix). Mounted at the top level so it's outside
	// the session-gated group.
	r.Mount("/v1", h.v1Router())

	// Native mobile (responder) app: token-authenticated read API. Mounted at
	// the top level, outside the session-gated group; CSRF-exempt (the
	// /mobile/ prefix is exempted in csrf.New).
	r.Mount("/mobile", h.mobileRouter())

	// API docs are public.
	r.Get("/openapi.json", h.openAPISpec)
	r.Get("/docs", h.docsPage)

	// Public pricing page — self-serve subscribers start here.
	r.Get("/pricing", h.pricingPage)

	// Public explainer — what backup drilling is and how Dokaz helps.
	r.Get("/how-it-works", h.howItWorks)

	// Onboarding sample dump — a known-good fixture a new user can run
	// their first drill against before connecting a real database.
	r.Get("/onboarding/sample.dump", h.onboardingSampleDump)

	r.Handle("/static/*", http.StripPrefix("/static/", http.FileServer(staticFS)))

	r.Get("/", h.index)

	// Auth endpoints get a tight per-IP rate limit on top of the login
	// throttle: the throttle protects one account, the limiter protects the
	// whole box from a credential-stuffing flood.
	r.Group(func(r chi.Router) {
		r.Use(h.authLimiter.Middleware(clientIPKey))
		r.Get("/login", h.loginPage)
		r.Post("/login", h.loginSubmit)
		r.Get("/signup", h.signupPage)
		r.Post("/signup", h.signupSubmit)
		// Second login step — driven by the mfa_pending session cookie, so
		// it sits with the other pre-auth routes under the per-IP limiter.
		r.Get("/login/mfa", h.mfaChallengePage)
		r.Post("/login/mfa", h.mfaChallengeSubmit)
		// Passwordless ("magic link") sign-in.
		r.Get("/login/magic", h.magicLinkRequestPage)
		r.Post("/login/magic", h.magicLinkRequest)
		r.Get("/login/magic/{token}", h.magicLinkConsume)
		// Social login (OAuth) — start + provider callback.
		r.Get("/auth/{provider}/start", h.oauthStart)
		r.Get("/auth/{provider}/callback", h.oauthCallback)
	})
	r.Post("/logout", h.logout)

	// Invitation accept page — public so the recipient can land before
	// signing up; POST requires login.
	r.Get("/invitations/{token}", h.invitationPage)
	r.Post("/invitations/{token}/accept", h.invitationAccept)

	// Email verification: the link from the signup email is public so it
	// works before the recipient signs in.
	r.Get("/verify-email/{token}", h.verifyEmail)

	// Legal + help pages are public.
	r.Get("/legal/terms", h.legalTerms)
	r.Get("/legal/privacy", h.legalPrivacy)
	r.Get("/legal/dpa", h.legalDPA)
	r.Get("/legal/subprocessors", h.legalSubprocessors)
	r.Get("/legal/cookies", h.legalCookies)
	r.Get("/help", h.helpPage)

	// Ending an impersonation: requires login but NOT staff — mid-
	// impersonation the effective user is the (non-staff) target.
	r.Group(func(r chi.Router) {
		r.Use(auth.RequireUser)
		r.Post("/impersonate/stop", h.impersonateStop)
	})

	// Staff SSO step-up — staff-gated but NOT behind the step-up itself,
	// otherwise reaching the verification page would be impossible.
	r.Group(func(r chi.Router) {
		r.Use(auth.RequireUser)
		r.Use(auth.RequireStaff)
		r.Get("/admin/sso", h.adminSSOPage)
		r.Get("/admin/sso/start", h.adminSSOStart)
	})

	// Staff admin panel — staff-gated, no account requirement (staff act
	// across accounts). requireStaffSSO also enforces a recent SSO step-up
	// when SSO is configured.
	r.Group(func(r chi.Router) {
		r.Use(auth.RequireUser)
		r.Use(h.requireStaffSSO)
		r.Get("/admin", h.adminHome)
		r.Get("/admin/users", h.adminUserSearch)
		r.Get("/admin/users/{id}", h.adminUserDetail)
		r.Post("/admin/users/{id}/impersonate", h.adminImpersonate)
		r.Get("/admin/drills/{id}", h.adminDrillDetail)
		r.Post("/admin/drills/{id}/replay", h.adminDrillReplay)
		r.Post("/admin/drills/{id}/regen-evidence", h.adminEvidenceRegen)
		r.Get("/admin/accounts/{id}", h.adminAccountDetail)
		r.Post("/admin/accounts/{id}/refund", h.adminRefund)
		r.Get("/admin/email", h.adminEmailReport)
	})

	r.Group(func(r chi.Router) {
		r.Use(auth.RequireUser)
		r.Use(auth.RequireAccount)
		// Authenticated traffic is rate-limited per account.
		r.Use(h.appLimiter.Middleware(accountKey))

		r.Get("/dashboard", h.dashboard)
		r.Post("/account/switch", h.accountSwitch)
		r.Post("/verify-email/resend", h.verifyEmailResend)

		// Two-factor auth — a per-user security setting, not RBAC-gated.
		r.Get("/account/mfa", h.mfaSetupPage)
		r.Post("/account/mfa/enable", h.mfaEnable)
		r.Post("/account/mfa/disable", h.mfaDisable)

		// Reads
		r.Group(func(r chi.Router) {
			r.Use(auth.RequireAction(auth.ActionTargetRead))
			r.Get("/databases", h.targetsList)
			r.Get("/databases/{id}", h.targetDetail)
		})
		r.Group(func(r chi.Router) {
			r.Use(auth.RequireAction(auth.ActionDrillRead))
			r.Get("/drills", h.drillsList)
			r.Get("/drills/{id}", h.drillDetail)
			r.Get("/drills/{id}/steps", h.drillStepsPartial)
			r.Get("/reports", h.reports)
			r.Get("/reports/export.csv", h.reportsExport)
		})
		r.Group(func(r chi.Router) {
			r.Use(auth.RequireAction(auth.ActionHeartbeatRead))
			r.Get("/heartbeats", h.heartbeatsList)
			r.Get("/heartbeats/{id}", h.heartbeatDetail)
			r.Get("/heartbeats/{id}/rows", h.heartbeatStatusPartial)
		})
		r.Group(func(r chi.Router) {
			r.Use(auth.RequireAction(auth.ActionEvidenceRead))
			r.Get("/drills/{id}/evidence", h.drillEvidence)
		})
		r.Group(func(r chi.Router) {
			r.Use(auth.RequireAction(auth.ActionAccountRead))
			r.Get("/account", h.accountSettings)
			r.Get("/account/audit", h.auditLog)
			r.Get("/account/webhooks", h.webhooksList)
			r.Get("/account/webhooks/{id}/deliveries", h.webhookDeliveries)
			r.Get("/account/api-keys", h.apiKeysList)
		})

		// Writes (RBAC-gated)
		r.Group(func(r chi.Router) {
			r.Use(auth.RequireAction(auth.ActionTargetWrite))
			r.Use(h.requireUnlapsedTrial)
			r.Get("/databases/new", h.targetNewPage)
			r.Post("/databases", h.targetCreate)
			r.Post("/databases/{id}/assertions", h.assertionCreate)
			r.Post("/databases/{id}/assertions/{assertion_id}/delete", h.assertionDelete)
			r.Post("/databases/{id}/schedule", h.targetScheduleUpdate)
		})
		r.Group(func(r chi.Router) {
			r.Use(auth.RequireAction(auth.ActionDrillWrite))
			r.Use(h.requireUnlapsedTrial)
			r.Post("/drills", h.drillCreate)
			// The free demo: run a drill against the built-in sample dataset.
			// Allowed on every plan (it's the sample, not a real backup).
			r.Post("/databases/sample-drill", h.runSampleDrill)
		})
		r.Group(func(r chi.Router) {
			r.Use(auth.RequireAction(auth.ActionHeartbeatWrite))
			r.Use(h.requireUnlapsedTrial)
			r.Get("/heartbeats/new", h.heartbeatNewPage)
			r.Post("/heartbeats", h.heartbeatCreate)
			r.Post("/heartbeats/{id}/pause", h.heartbeatPause)
			r.Post("/heartbeats/{id}/resume", h.heartbeatResume)
			r.Post("/heartbeats/{id}/delete", h.heartbeatDelete)
		})
		r.Group(func(r chi.Router) {
			r.Use(auth.RequireAction(auth.ActionMemberWrite))
			r.Post("/account/invitations", h.inviteCreate)
			r.Post("/account/members/{user_id}", h.memberUpdate)
			r.Post("/account/members/{user_id}/remove", h.memberRemove)
			r.Post("/account/members/{user_id}/transfer-ownership", h.memberTransferOwnership)
		})
		// Webhook management + compliance actions are account-write
		// concerns. The delete handler additionally requires owner.
		r.Group(func(r chi.Router) {
			r.Use(auth.RequireAction(auth.ActionAccountWrite))
			r.Post("/account/webhooks", h.webhookCreate)
			r.Post("/account/webhooks/{id}/delete", h.webhookDelete)
			r.Post("/account/webhooks/{id}/deliveries/{delivery_id}/replay", h.webhookReplay)
			r.Post("/account/api-keys", h.apiKeyCreate)
			r.Post("/account/api-keys/{id}/revoke", h.apiKeyRevoke)
			r.Get("/account/export", h.accountExport)
			r.Post("/account/delete", h.accountDelete)
		})
		// Billing — starting Checkout / opening the Customer Portal.
		r.Group(func(r chi.Router) {
			r.Use(auth.RequireAction(auth.ActionBillingWrite))
			r.Post("/account/billing/checkout", h.billingCheckout)
			r.Post("/account/billing/portal", h.billingPortal)
		})
	})

	return r
}

// metricsHandler serves Prometheus metrics. When METRICS_TOKEN is set it
// requires a matching bearer token; unset leaves /metrics open for local
// dev. Production should set the token (or scrape over a private network).
func (h *Handlers) metricsHandler() http.Handler {
	inner := h.obs.Metrics.Handler()
	if h.metricsToken == "" {
		return inner
	}
	want := "Bearer " + h.metricsToken
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		if subtle.ConstantTimeCompare([]byte(r.Header.Get("Authorization")), []byte(want)) != 1 {
			w.Header().Set("WWW-Authenticate", "Bearer")
			http.Error(w, "unauthorized", http.StatusUnauthorized)
			return
		}
		inner.ServeHTTP(w, r)
	})
}

// stampAccountForLogs enriches the request-scoped log fields with the
// resolved account ID, once LoadCurrentAccount has run.
func stampAccountForLogs(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		if a, ok := auth.CurrentAccountFromContext(r.Context()); ok {
			r = r.WithContext(obs.WithAccountID(r.Context(), a.ID.String()))
		}
		next.ServeHTTP(w, r)
	})
}

// clientIPKey buckets the rate limiter by client IP.
func clientIPKey(r *http.Request) string {
	return "ip:" + audit.ClientIP(r)
}

// accountKey buckets by current account, falling back to IP when somehow
// account context is missing (shouldn't happen behind RequireAccount).
func accountKey(r *http.Request) string {
	if a, ok := auth.CurrentAccountFromContext(r.Context()); ok {
		return "acct:" + a.ID.String()
	}
	return "ip:" + audit.ClientIP(r)
}

// securityHeaders sets a baseline set of headers. CSP is intentionally strict;
// inline scripts are not allowed. Tighten further per route as needed.
// maxRequestBody caps the request body for every route. Form handlers call
// r.ParseForm(), which buffers the whole body into memory, so without a cap an
// unauthenticated multi-gigabyte POST is a trivial memory-exhaustion DoS. JSON
// handlers apply their own (smaller) io.LimitReader on top of this.
const maxRequestBody = 2 << 20 // 2 MiB

func limitRequestBody(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		if r.Body != nil {
			r.Body = http.MaxBytesReader(w, r.Body, maxRequestBody)
		}
		next.ServeHTTP(w, r)
	})
}

func securityHeaders(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		h := w.Header()
		h.Set("Content-Security-Policy",
			"default-src 'self'; "+
				"script-src 'self'; "+
				"style-src 'self'; "+
				"img-src 'self' data:; "+
				"font-src 'self'; "+
				"connect-src 'self'; "+
				"frame-ancestors 'none'; "+
				"base-uri 'self'; "+
				"form-action 'self'")
		h.Set("Referrer-Policy", "strict-origin-when-cross-origin")
		h.Set("X-Content-Type-Options", "nosniff")
		h.Set("X-Frame-Options", "DENY")
		h.Set("Permissions-Policy", "camera=(), microphone=(), geolocation=()")
		// HSTS is set unconditionally; harmless on HTTP, mandatory on HTTPS.
		h.Set("Strict-Transport-Security", "max-age=31536000; includeSubDomains; preload")
		next.ServeHTTP(w, r)
	})
}